fixes in team meeting

+++ src/main/java/dgb/GradebookService.java
@@ -244,6 +244,9 @@ public class GradebookService extends RestTemplate {
 	// DELETE
 	public void deleteSecondaryGradebook (Integer gradebookId) throws GradebookNotFoundException {
 		Gradebook gradebook = getGradebookById(gradebookId);
+		if (gradebook.getIsPrimaryServer()) {
+			throw new SecondaryEditNotAllowedException("Something");
+		}
 		gradebookRepo.deleteById(gradebookId);
 	}
